+++ SBBMobileDesignSystemSwiftUIDemo/Views/DetailViews/IconDemo.swift
@@ -100,7 +100,7 @@ struct IconDemo: View {
                 .padding(16)
         }
             .navigationBarTitle("Icons")
-            .background(Color.sbbColor(.background).edgesIgnoringSafeArea(.bottom))
+            .sbbStyle()
             .colorScheme(colorScheme)
     }
 }