+++ SBBMobileDesignSystemSwiftUIDemo/Views/DetailViews/LoadingIndicatorDemo.swift
@@ -60,7 +60,7 @@ struct LoadingIndicatorDemo: View {
                 .padding(16)
         }
         .navigationBarTitle("LoadingIndicator")
-        .background(Color.sbbColor(.background).edgesIgnoringSafeArea(.bottom))
+        .sbbStyle()
         .colorScheme(colorScheme)
     }
 }