+++ SBBMobileDesignSystemSwiftUIDemo/Views/DetailViews/ModalViewDemo.swift
@@ -123,7 +123,7 @@ struct ModalViewDemo: View {
         }
             .padding(16)
             .navigationBarTitle("ModalView")
-            .background(Color.sbbColor(.background).edgesIgnoringSafeArea(.bottom))
+            .sbbStyle()
             .colorScheme(colorScheme)
     }
 }